background test

+++ GAME1017_Template_W01/Background.h
@@ -2,7 +2,7 @@
 #include "Sprite.h"
 class Background : public Sprite {
 public:
-	Background(SDL_Rect s, SDL_FRect d, const char* p, std::string k, int velocity) : Sprite(s, d, p, k) 
+	Background(SDL_Rect s, SDL_FRect d, const char* p, std::string k, float velocity) : Sprite(s, d, p, k) 
 	{
 		m_velocity = velocity;
 	}
